Feeling sleepy, will probably miss something. Got model fits to run but they look crazy, just like Gaussian fits.
Doesn't seem to depend on initial conditions, or at least I get effectivly the same result when I use the position of max intensity or I use the true molecule location as initial guess. Something is wrong, but I dont know what. What would cause the odel to give *almost* the same result as the gaussian localization?

I need to pull a Kevin and just lot the fit result over the data and see if they actually line up. How can that be? are my model parameters wack? I don;t think my fluorophore is tuned properly but that shouldn't matter. Of well its time to give up for tonight.